+++ libs/hscim/src/Web/Scim/Server.hs
@@ -25,6 +25,7 @@ module Web.Scim.Server
 
     -- * API tree
     SiteAPI,
+    Site (..),
     siteServer,
 
     -- ** API subtrees, useful for tests