+++ CHANGELOG.md
@@ -31,16 +31,27 @@
 
 ## API Changes
 
+* Add `POST /conversations/list-ids` (#1686)
+* Deprecate `GET /converstations/ids` (#1686)
+
 ## Features
+
+* Client functions for the hscim library (#1699)
 
 ## Bug fixes and other updates
 
+## Federation changes (alpha feature, do not use yet)
+
+* Add new API to list paginated qualified conversation ids (#1686)
+
 ## Documentation
 
-*  fix swagger: mark name in UserUpdate as optional (#1691)
+* fix swagger: mark name in UserUpdate as optional (#1691)
 
 ## Internal changes
 
+* Replaced uses of `UVerb` and `EmptyResult` with `MultiVerb` (#1693)
+* Added a mechanism to derive `AsUnion` instances automatically (#1693)
 
 # [2021-08-02]